+++ HISTORY.md
@@ -2,6 +2,7 @@
 
 ## API Changes
 * Use `bool` instead of `uint8` for Boolean dtype in `dataframe_.py` [#1154](https://github.com/TileDB-Inc/TileDB-Py/pull/1154)
+* Support QueryCondition OR operator [#1146](https://github.com/TileDB-Inc/TileDB-Py/pull/1146)
 
 # TileDB-Py 0.15.3 Release Notes